test: 테스트를 적절한 위치로 이동
